Update export logo.svg to avoid VS C2026 bug
Visual Studio enforces the length of the string in a way that gcc and clang for linuxbsd and macos platforms do not.  (https://learn.microsoft.com/en-us/cpp/error-messages/compiler-errors-1/compiler-error-c2026?view=msvc-170)

The logo.svg previously used in the exports were converted from png raster images.  This made the resulting svg file rather large.  This patch replaces the export logo.svg with a simple black Tekisasu graphic.

There's a way to resolve this by changing how the logo_svn.gen.h is generated, but I'd rather not alter upstream Godot in such a significant way that does not have any effect on released games, as this results in an overly complicated patching process with every rebase.